requirements.txt
python-dotenv
netmiko